# AISEA.ApiService/AISEA.ApiService.WebApi/OpenAPI/curriculum-combo-api.yaml
openapi: 3.0.3
info:
  title: AISEA Curriculum and Subject Combinations API
  description: |
    RESTful API system for managing curriculum and subject combinations in the AISEA academic system.
    
    ## Features
    - Complete CRUD operations for curricula and subject combinations
    - Version tracking for curriculum changes
    - Subject prerequisite validation
    - Student enrollment management
    - Advanced search and filtering capabilities
    - Role-based access control
    
    ## Authentication
    All endpoints require JWT authentication via Bearer token in the Authorization header.
    Some endpoints have additional role-based restrictions.
    
    ## Error Handling
    The API returns consistent error responses with appropriate HTTP status codes:
    - 400: Bad Request (validation errors)
    - 401: Unauthorized (invalid/missing token)
    - 403: Forbidden (insufficient permissions)
    - 404: Not Found (resource doesn't exist)
    - 500: Internal Server Error
    
  version: 1.0.0
  contact:
    name: AISEA Development Team
    email: dev@aisea.edu
servers:
  - url: http://localhost:5000/api
    description: Development server
  - url: https://api.aisea.edu/api
    description: Production server

security:
  - BearerAuth: []

paths:
  # Curriculum Management Endpoints
  /curriculum:
    get:
      tags:
        - Curriculum
      summary: Get paginated list of curricula
      description: Retrieve curricula with search, filtering, and pagination support
      parameters:
        - name: pageNumber
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
        - name: search
          in: query
          description: Search in curriculum name, code, or program name
          schema:
            type: string
        - name: programId
          in: query
          description: Filter by program ID
          schema:
            type: integer
            format: int64
        - name: effectiveDateFrom
          in: query
          description: Filter curricula effective from this date
          schema:
            type: string
            format: date-time
        - name: effectiveDateTo
          in: query
          description: Filter curricula effective until this date
          schema:
            type: string
            format: date-time
        - name: isActive
          in: query
          description: Filter by active status (effective date <= now)
          schema:
            type: boolean
        - name: sortBy
          in: query
          description: Sort field
          schema:
            type: string
            enum: [curriculumName, curriculumCode, effectiveDate, programName]
            default: curriculumName
        - name: sortOrder
          in: query
          description: Sort order
          schema:
            type: string
            enum: [asc, desc]
            default: asc
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedCurriculumResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Curriculum
      summary: Create a new curriculum
      description: Create a new curriculum with subjects and automatic version tracking
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCurriculumRequest'
            example:
              programId: 1
              curriculumCode: "SE2024"
              curriculumName: "Software Engineering 2024"
              effectiveDate: "2024-09-01T00:00:00Z"
              description: "Updated curriculum for Software Engineering program"
              subjects:
                - subjectId: 1
                  semesterNumber: 1
                  isMandatory: true
                - subjectId: 2
                  semesterNumber: 1
                  isMandatory: false
      responses:
        '200':
          description: Curriculum created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Curriculum created successfully."
                  curriculumId:
                    type: integer
                    format: int64
                    example: 123
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

  /curriculum/{id}:
    get:
      tags:
        - Curriculum
      summary: Get curriculum details
      description: Get detailed curriculum information including subjects and version history
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurriculumDetailResponse'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags:
        - Curriculum
      summary: Update curriculum
      description: Update curriculum and create new version if significant changes are made
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCurriculumRequest'
      responses:
        '200':
          description: Curriculum updated successfully
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - Curriculum
      summary: Delete curriculum
      description: Soft delete a curriculum (Admin only)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Curriculum deleted successfully
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  /curriculum/active:
    get:
      tags:
        - Curriculum
      summary: Get active curricula
      description: Get simplified list of all active curricula
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CurriculumResponse'

  # Subject Combinations (Combo) Endpoints
  /combo:
    get:
      tags:
        - Subject Combinations
      summary: Get paginated list of subject combinations
      description: Retrieve subject combinations with filtering and pagination
      parameters:
        - name: pageNumber
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
        - name: search
          in: query
          description: Search in combo name or description
          schema:
            type: string
        - name: programId
          in: query
          description: Filter by program ID
          schema:
            type: integer
            format: int64
        - name: semesterNumber
          in: query
          description: Filter by semester number
          schema:
            type: integer
        - name: difficultyLevel
          in: query
          description: Filter by difficulty level
          schema:
            type: string
            enum: [Easy, Medium, Hard]
        - name: isAvailable
          in: query
          description: Filter by availability (has open slots)
          schema:
            type: boolean
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [comboName, semesterNumber, difficultyLevel, maxStudents]
            default: comboName
        - name: sortOrder
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: asc
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedComboResponse'

    post:
      tags:
        - Subject Combinations
      summary: Create a new subject combination
      description: Create a new subject combination with prerequisite validation
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateComboRequest'
            example:
              comboName: "Advanced Programming Track"
              comboDescription: "Intensive programming courses for advanced students"
              semesterNumber: 3
              programId: 1
              difficultyLevel: "Hard"
              maxStudents: 30
              subjectIds: [1, 2, 3]
              prerequisites:
                - subjectId: 4
                  isRequired: true
                - subjectId: 5
                  isRequired: false
      responses:
        '200':
          description: Subject combination created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  comboId:
                    type: integer
                    format: int64

  /combo/{id}:
    get:
      tags:
        - Subject Combinations
      summary: Get subject combination details
      description: Get detailed information about a subject combination including enrollments
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComboDetailResponse'

    put:
      tags:
        - Subject Combinations
      summary: Update subject combination
      description: Update an existing subject combination
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateComboRequest'
      responses:
        '200':
          description: Subject combination updated successfully

    delete:
      tags:
        - Subject Combinations
      summary: Delete subject combination
      description: Delete a subject combination (only if no active enrollments)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Subject combination deleted successfully
        '400':
          description: Cannot delete combo with active enrollments

  /combo/{id}/availability:
    get:
      tags:
        - Subject Combinations
      summary: Check combo availability for student
      description: Check if the current student can enroll in a subject combination
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Availability check result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComboAvailabilityResponse'
              example:
                comboId: 123
                comboName: "Advanced Programming Track"
                isAvailable: false
                availableSlots: 0
                unavailableReasons:
                  - "Combo is at full capacity"
                  - "Missing required prerequisites"
                missingPrerequisites:
                  - "Data Structures and Algorithms"
                  - "Object-Oriented Programming"

  # Student Enrollment Endpoints
  /combo/enroll:
    post:
      tags:
        - Student Enrollment
      summary: Enroll student in subject combination
      description: Enroll a single student in a subject combination with prerequisite validation
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StudentEnrollmentRequest'
            example:
              comboId: 123
              studentId: 456
              notes: "Student requested this combination"
      responses:
        '200':
          description: Student enrolled successfully
        '400':
          description: Enrollment failed (capacity, prerequisites, etc.)

  /combo/enroll/bulk:
    post:
      tags:
        - Student Enrollment
      summary: Bulk enroll students
      description: Enroll multiple students in a subject combination
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BulkEnrollmentRequest'
            example:
              comboId: 123
              studentIds: [456, 789, 101]
              notes: "Batch enrollment for advanced track"
      responses:
        '200':
          description: Students enrolled successfully

  /combo/{comboId}/students/{studentId}:
    delete:
      tags:
        - Student Enrollment
      summary: Unenroll student from subject combination
      description: Remove a student from a subject combination
      security:
        - BearerAuth: []
      parameters:
        - name: comboId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: studentId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Student unenrolled successfully
        '404':
          description: Enrollment not found

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    # Curriculum Schemas
    CreateCurriculumRequest:
      type: object
      required:
        - programId
        - curriculumCode
        - curriculumName
        - effectiveDate
      properties:
        programId:
          type: integer
          format: int64
          description: ID of the associated program
        curriculumCode:
          type: string
          maxLength: 50
          description: Unique curriculum code
          example: "SE2024"
        curriculumName:
          type: string
          maxLength: 255
          description: Human-readable curriculum name
          example: "Software Engineering 2024"
        effectiveDate:
          type: string
          format: date-time
          description: Date when curriculum becomes effective
        description:
          type: string
          description: Optional curriculum description
        subjects:
          type: array
          items:
            $ref: '#/components/schemas/CurriculumSubjectRequest'

    UpdateCurriculumRequest:
      type: object
      required:
        - curriculumName
        - effectiveDate
      properties:
        curriculumName:
          type: string
          maxLength: 255
        effectiveDate:
          type: string
          format: date-time
        description:
          type: string
        subjects:
          type: array
          items:
            $ref: '#/components/schemas/CurriculumSubjectRequest'

    CurriculumSubjectRequest:
      type: object
      required:
        - subjectId
        - semesterNumber
        - isMandatory
      properties:
        subjectId:
          type: integer
          format: int64
        semesterNumber:
          type: integer
          minimum: 1
          maximum: 8
        isMandatory:
          type: boolean

    CurriculumResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        programId:
          type: integer
          format: int64
        programName:
          type: string
        programCode:
          type: string
        curriculumCode:
          type: string
        curriculumName:
          type: string
        effectiveDate:
          type: string
          format: date-time
        description:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        totalSubjects:
          type: integer
        totalCredits:
          type: integer

    CurriculumDetailResponse:
      allOf:
        - $ref: '#/components/schemas/CurriculumResponse'
        - type: object
          properties:
            subjects:
              type: array
              items:
                $ref: '#/components/schemas/CurriculumSubjectResponse'
            versions:
              type: array
              items:
                $ref: '#/components/schemas/CurriculumVersionResponse'

    CurriculumSubjectResponse:
      type: object
      properties:
        subjectId:
          type: integer
          format: int64
        subjectCode:
          type: string
        subjectName:
          type: string
        credits:
          type: integer
        semesterNumber:
          type: integer
        isMandatory:
          type: boolean
        prerequisites:
          type: array
          items:
            type: string

    CurriculumVersionResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        version:
          type: string
        effectiveDate:
          type: string
          format: date-time
        changeDescription:
          type: string
        createdAt:
          type: string
          format: date-time
        createdBy:
          type: string

    PagedCurriculumResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/CurriculumResponse'
        totalCount:
          type: integer
        pageNumber:
          type: integer
        pageSize:
          type: integer

    # Subject Combination Schemas
    CreateComboRequest:
      type: object
      required:
        - comboName
        - semesterNumber
        - programId
        - difficultyLevel
        - maxStudents
        - subjectIds
      properties:
        comboName:
          type: string
          maxLength: 255
        comboDescription:
          type: string
        semesterNumber:
          type: integer
          minimum: 1
          maximum: 8
        programId:
          type: integer
          format: int64
        difficultyLevel:
          type: string
          enum: [Easy, Medium, Hard]
        maxStudents:
          type: integer
          minimum: 1
        subjectIds:
          type: array
          items:
            type: integer
            format: int64
          minItems: 1
        prerequisites:
          type: array
          items:
            $ref: '#/components/schemas/ComboPrerequisiteRequest'

    UpdateComboRequest:
      type: object
      required:
        - comboName
        - semesterNumber
        - difficultyLevel
        - maxStudents
        - subjectIds
      properties:
        comboName:
          type: string
          maxLength: 255
        comboDescription:
          type: string
        semesterNumber:
          type: integer
          minimum: 1
          maximum: 8
        difficultyLevel:
          type: string
          enum: [Easy, Medium, Hard]
        maxStudents:
          type: integer
          minimum: 1
        subjectIds:
          type: array
          items:
            type: integer
            format: int64
          minItems: 1
        prerequisites:
          type: array
          items:
            $ref: '#/components/schemas/ComboPrerequisiteRequest'

    ComboPrerequisiteRequest:
      type: object
      required:
        - subjectId
      properties:
        subjectId:
          type: integer
          format: int64
        isRequired:
          type: boolean
          default: true

    ComboResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        comboName:
          type: string
        comboDescription:
          type: string
        semesterNumber:
          type: integer
        programId:
          type: integer
          format: int64
        programName:
          type: string
        difficultyLevel:
          type: string
        maxStudents:
          type: integer
        currentEnrollment:
          type: integer
        isAvailable:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        subjects:
          type: array
          items:
            $ref: '#/components/schemas/ComboSubjectResponse'

    ComboDetailResponse:
      allOf:
        - $ref: '#/components/schemas/ComboResponse'
        - type: object
          properties:
            prerequisites:
              type: array
              items:
                $ref: '#/components/schemas/ComboPrerequisiteResponse'
            enrollments:
              type: array
              items:
                $ref: '#/components/schemas/StudentEnrollmentResponse'
            statistics:
              $ref: '#/components/schemas/ComboStatisticsResponse'

    ComboSubjectResponse:
      type: object
      properties:
        subjectId:
          type: integer
          format: int64
        subjectCode:
          type: string
        subjectName:
          type: string
        credits:
          type: integer
        description:
          type: string

    ComboPrerequisiteResponse:
      type: object
      properties:
        subjectId:
          type: integer
          format: int64
        subjectCode:
          type: string
        subjectName:
          type: string
        isRequired:
          type: boolean

    ComboAvailabilityResponse:
      type: object
      properties:
        comboId:
          type: integer
          format: int64
        comboName:
          type: string
        isAvailable:
          type: boolean
        availableSlots:
          type: integer
        unavailableReasons:
          type: array
          items:
            type: string
        missingPrerequisites:
          type: array
          items:
            type: string

    PagedComboResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ComboResponse'
        totalCount:
          type: integer
        pageNumber:
          type: integer
        pageSize:
          type: integer

    # Student Enrollment Schemas
    StudentEnrollmentRequest:
      type: object
      required:
        - comboId
        - studentId
      properties:
        comboId:
          type: integer
          format: int64
        studentId:
          type: integer
          format: int64
        notes:
          type: string

    BulkEnrollmentRequest:
      type: object
      required:
        - comboId
        - studentIds
      properties:
        comboId:
          type: integer
          format: int64
        studentIds:
          type: array
          items:
            type: integer
            format: int64
          minItems: 1
        notes:
          type: string

    StudentEnrollmentResponse:
      type: object
      properties:
        studentId:
          type: integer
          format: int64
        studentCode:
          type: string
        studentName:
          type: string
        enrolledAt:
          type: string
          format: date-time
        notes:
          type: string
        status:
          type: string
          enum: [Active, Completed, Dropped, Withdrawn]

    ComboStatisticsResponse:
      type: object
      properties:
        totalCredits:
          type: integer
        averageGPA:
          type: number
          format: double
        completionRate:
          type: integer
          description: Percentage of students who completed the combo
        gradeDistribution:
          type: object
          additionalProperties:
            type: integer
          example:
            A: 5
            B: 10
            C: 8
            D: 2
            F: 1

    # Common Error Response
    ErrorResponse:
      type: object
      properties:
        status:
          type: integer
        message:
          type: string
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string

  responses:
    BadRequest:
      description: Bad request - validation errors
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            status: 400
            message: "Validation failed"
            errors:
              curriculumCode: ["Curriculum code is required"]
              effectiveDate: ["Effective date must be in the future"]

    Unauthorized:
      description: Unauthorized - invalid or missing authentication token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            status: 401
            message: "Unauthorized access"

    Forbidden:
      description: Forbidden - insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            status: 403
            message: "Insufficient permissions to perform this action"

    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            status: 404
            message: "Curriculum not found"

tags:
  - name: Curriculum
    description: Curriculum management operations
  - name: Subject Combinations
    description: Subject combination (combo) management operations
  - name: Student Enrollment
    description: Student enrollment in subject combinations